fix(security): split pure authorization logic into separate module for Node 18 tests
Node 18's vitest runtime fails to load jwks-rsa (via transitive jose
ESM-only dep) when the test imports anything from user-token-validator.
Extract the pure authorizeUserClaims function + its types into a new
src/user-token-authorization.ts with no JWT/JWKS imports, so the test
pulls only the policy layer. user-token-validator.ts re-exports the
public surface, keeping existing callers (http-server.ts) unchanged.
